Change return type of get_power_unit function
This changes the return type of the get_power_unit function to conform
to the same standard as other get_xxx() functions. This makes it
runnable using the same code as is used for the other parallellizable
functions.

An extra column 'db_index' is added. This should be useful in
reconstructing which database indices were downloaded already and which
have failed.