Correct types in EEG tables
Too many columns where specified as available for all technologies.

Those columns are now moved to the respective tables, where data is
available for these columns. This requires to remove two columns from to
tables of the database that is already filling up.

The columns

- PrototypAnlage
- PilotAnlage

need to be removed from

- solar_eeg
- biomass_eeg