Fix indentation for collection additional data requested
The wrong indentation for data types eeg, kwk and permit was observed
while querying a single unit with an not empty database. It led to an
"reference before assigned error" for basic_unit.
The wrong indentation (prior to this fix) led to missing additional data
requests: without having the data types eeg, kwk and permit inside the
for-loop, only for the last unit additional data gets requested.

This might provide a fix for #154.