Replace katalogwerte in cleansed database
In the raw data, many entries are only given as numbers. Their meaning
can be read from the xml file katalogwerte. In the cleansed data, these
numbers are replaced by their original meaning. The table katalogkategorien
is not used anymore. Instead, a hardcoded list "columns_to_replace.py" is
created with the names of all columns were one has to replace the values.