Modify function flatten_dict
Adjust the function flatten_dict to solve an IndexError.

The function raised an 'IndexError: list index out of range', if a data point was processed that did not contain any 'VerknuepfteEinheiten'. The function tried to get the first, in this case non-existent, MaStRNummer and assign it as a new value to 'VerknuepfteEinheiten' (flatten dict). This led to the error.

The function was modified to check for empty lists for the key 'VerknuepfteEinheiten' and pass over an empty list in case of emptiness. Hence, the data point will have no 'VerknuepfteEinheiten'.