Add pytest fixture decorator
With the decorator, the connection con does not need to be executed in each function, but it is passed as an argument.
scope="module" makes the con function run only once at the beginning.
After the con parameter is yielded, the connection is closed in the second function call (at the end).

[Ticket: #202]