Refactor function initialize_db as method from Mastr class
The method to initialize the structure of the database has to be called
for both the API and the bulk download. It was therefore moved to a method
of the Mastr class. It is now called when a Mastr object is initialized.
Additionally, python black was used for code formatting.

[Ticket: #202]