Adapt ORM for bulk download
The bulk download adds new tables, which are added to orm.py
Those tables have only their name and the primary key defined. A description of
all columns is added in future.
Second, the dictionary tablename_mapping is introduced.
It maps the table names from the bulk download to the names from the API
download.

[Ticket: 202]