Correct sqlite writing bug
If the table is written again to sql, it first has to
be deleted (otherwise the not unique constraint does not allow
to add new elements).
Additionally, the function write_single_entries_until_not_unique_comes_up
was created, since there appear entries with the same primary key in the
xml data. These entries are simply deleted.

[Ticket: #202]