Use the parameter technology instead of bulk_include_tables
The selection of tables to be created/filled happens for bulk download also by technology. With a further function the technologies are mapped to the .xml file names.
This way the parameter of mastr.download() 'bulk_include_tables' is avoided and the function usage is standardized for bulk and API.