Add stream_results=True to sqlite engine #269
The option will enable real batching and allow to only load rows from the database when the user requests them. This feature was introduced to allow the data export in chunks without loading all data into the memory. SQLAlchemy is using a client-side cursor: Without stream_results=True it loads all the data into memory, and then hands the Pandas API x-chunksize rows at a time. This could still resolve in memory issue if the data set is big enough - thus, this implementation.